all tensors should be same dtype

+++ MMAA/calc_test_loss.py
@@ -72,14 +72,20 @@ if __name__ == "__main__":
                                 if modality == "fmri":  
                                     continue                    
                                     
-                                X_test_tensor = torch.from_numpy(getattr(X_test, f"{modality}_data"), dtype=torch.double)
-                                X_train_tensor = torch.from_numpy(getattr(X_train, f"{modality}_data"),dtype=torch.double)
+                                X_test_tensor = torch.from_numpy(getattr(X_test, f"{modality}_data"))
+                                X_train_tensor = torch.from_numpy(getattr(X_train, f"{modality}_data"))
 
                                 # C is the same for all subjects and modalities
                                 C = torch.from_numpy(C)
                                 # S is unique for each subject and modality
                                 S = torch.from_numpy(S[idx,:,:])
 
+                                #make all  tensors double
+                                X_test_tensor = X_test_tensor.double()
+                                X_train_tensor = X_train_tensor.double()
+                                C = C.double()
+                                S = S.double()
+                                
                                 print("debugging",file=sys.stderr)
                                 # print the current torch dtype
                                 print(f"X_test_tensor dtype: {X_test_tensor.dtype}",file=sys.stderr)
